requirements.txt
pycryptodome
PyPDF2
pdf2image
Pillow
tk